Introduce Epoch
For calculations such a precession, the epoch is one parameter to
describe equatorial coordinates.

In this library, an epoch is expressed as a Julian day number. This is
especially useful as it is in this form it's used in many calculations.

Some standard values are pre-defined, such as J2000 or B1950, but it is
possible to set one from a Time object. `DateTime.ajd` does the rest of
the job.

This commit also adds the epoch as an attribute for equatorial
coordinates, while it is set as J2000 by default when not explicitly
set.