Add Angle::as_hms
Before, we had to convert angles in hours into decimal values to set up
coordinates. This was a bit annoying as most of the time the coordinates
as specified in HMS format, so the test example often differed from the
books.

Now, it is possible to create an angle in hours from HMS values. Test
examples are closer to what's written in source books.