Add Angle::as_dms
Before, we had to convert angles in degrees into decimal values to set
up coordinates. This was a bit annoying as most of the time the
coordinates are specified in DMS format, so the test examples often
differed from the books.

Now, it is possible to create an angle in degrees from DMS values. Test
examples are closer to what's in source books.