Consider latitude and longitude as angles
Before, latitude and longitude were often considered as plain values,
implied in degrees. A few conversions were done with magic numbers such
as "* 15" or "/ 15" when dealing with angles in hours.

Now, latitude and longitude are always defined as Angle objects. This
makes it easier to manipulate them and convert them without introducing
magic numbers. It also goes with the philosophy of the library where
most things are objects.

As a bonus, this enabled to remove to legacy code.

Fixed #8