Add Aberration computation
This introduces `Astronoby::Aberration` to compute apparent ecliptic
coordinates from true ecliptic coordinates with correction for
aberration, the effect due to the speed of the Earth in its orbit around
the Sun.

This implementation will need to be improved in the future to take an
epoch as parameter instead of the Sun's ecliptic longitude, so that this
value to automatically computed within the algorithm, from the given
epoch.