Improve Aberration ease of use
Before, we had to pass the Sun's ecliptic longitude as parameter to
compute the aberration correction. It was not ideal as it created a
dependency with the Sun body everywhere the aberration was needed, while
the Sun's ecliptic longitude is just an internal value for the formula.

Now, the Aberration expects an epoch and computes the Sun's ecliptic
longitude internally.