Normalize angles to a single circle
Before, it was possible to instantiate degrees with values above a full
circle, like 365 degrees, 3×π or 25 hours.

While this is mathematically acceptable, it makes it ambiguous when we
will add the ability to compare angles between each other.

For example, on a single circle, 3×π is a smaller angle than π÷2.

Without normalization, the angle's value is just a number that can go up
and down to the infinity.

This introduces angle normalization, so that angles are always included
in the following ranges (depending on their unit):
- [-2π, 2π] radians
- [0, 360] degrees
- [0, 24] hours

Fixes #19