release: Bump version to 0.1.0
Releasing the 0.1.0 version of the gem.

This is the first official release. Good practices and standards have
not been respected so far, as I mainly added features and refactored the
code with no concern with backward compatibility.

Now that the gem is going to be available on RubyGems.org, I aim to be
better at this and enable potential users to use this gem safely.

* Support angles in hours
* Support coordinates
  * `Astronoby::Coordinates::Horizontal`
  * `Astronoby::Coordinates::Equatorial`
  * `Astronoby::Coordinates::Ecliptic`
* Add new body `Astronoby::Sun`
* Add `Astronoby::Aberration`
* Add `Astronoby::Epoch`
* Add `Astronoby::MeanObliquity`
* Add `Astronoby::TrueObliquity`
* Add `Astronoby::Nutation`
* Add `Astronoby::Precession`
* Add `Astronoby::Refraction`
* Add utils
  * `Astronoby::Util::Astrodynamics`
  * `Astronoby::Util::Time`
  * `Astronoby::Util::Trigonometry`