Change method to calculate rising and setting times
Before, the rising and setting times of a punctual body (mostly stars)
were computed based on formulas from Celestial Calculations.

Now, they are computed based on formulas from Practical Astronomy with
your Calculator or Spreadsheet. These formulas feel a bit more natural
to me and allow for an easy add of vertical shift angle.

At the moment, the vertical shift is either ignored, or a fixed value of
34 arc minutes for the atmospheric refraction.