Add #distance and #angular_size to Sun
With the Sun's true anomaly available, it is convenient to compute the
Earth-Sun distance and the Sun's apparent angular size.

The distance is of course an approximation. It is returned in meters as
it is a SI unit, but it must not be seen as a value precise by the
meter. For example, it doesn't count the observer's elevation or height.

Note that in the tests added, the precision for these two values taken
from books examples weren't as high as in the tests. Therefore, the
reader won't find matching results from the books. Also, these values
will likely be updated as the precision of the library increases, which
is acceptable as they weren't that close to the book examples in the
first place.