Accurate setting and rising times for punctual bodies (#29)
This changes the method to compute setting and rising times for punctual bodies in order to achieve better precision.

This new method is based on _Practical Astronomy with your Calculator or Spreadsheet_ as it makes it easier to add a vertical shift component and reach more accurate apparent coordinates.

This profoundly changes `Astronoby::Refraction` class in naming and flexibility. It now allows to either return a refraction angle, or corrected horizontal coordinates.

A new `Astronoby::Observer` class is introduced to represent an observer and several component based on their location such as the temperature, elevation and atmospheric pressure, if such information are available.

For now, `Astronoby::Body` doesn't benefit much from these new changes. The class will probably be refactored at some point depending on the features that will be developed around rising and setting times for different bodies such as stars, planets, etc.