Add #distance and #angular_size to Sun (#30)
With the Sun's true anomaly available, it is convenient to compute the Earth-Sun distance and the Sun's apparent angular size.

The distance is of course an approximation. It is returned in meters as it is a SI unit, but it must not be seen as a value precise by the meter. For example, it doesn't count the observer's elevation or height.

Note that in the tests added, the precision for these two values taken from books examples weren't as high as in the tests. Therefore, the reader won't find matching results from the books. Also, these values will likely be updated as the precision of the library increases, which is acceptable as they weren't that close to the book examples in the first place.

```rb
time = Time.utc(2010, 5, 6)
epoch = Astronoby::Epoch.from_time(time)
sun = described_class.new(epoch: epoch)

sun.earth_distance.round
# => 150902254024

sun.angular_size.str(:dms)
# => "+0° 31′ 42.6789″"
```