Compute the equation of time
The equation of time is the difference of the real Sun time and the mean
Sun time. This difference is due to the elliptical shape of the Earth's
orbit, and the obliquity of the orbit.

`Sun` now has a `#equation_of_time` method that returns an integer: the
time difference in seconds.

```rb
date = Date.new(2010, 7, 27)

Astronoby::Sun.equation_of_time(date: date)
```