Rename Angle::as_* into Angle::from_*
The initial names `as_radians`, `as_degrees` and `as_hours` were
misleading since we made `Angle` a single class for handling all units
of angles, always stored in radians.

It use to make sense when we had classes like `Radian`, `Degree` and
`Hour`, but now that an angle is always stored in radians, `as_*` could
let the reader think that the angle is going to be initialized in the
specified unit.

These methods have been renamed into the following:
* `::from_radians`
* `::from_degrees`
* `::from_hours`
* `::from_dms`
* `::from_hms`