Drop BigDecimal
`BigDecimal` was introduced early in the project from my belief that
high accuracy was needed.

It turns out that the number of digits in floats are enough to keep a
precision up to 0.0001 arcseconds in most calculations.

Using `BigDecimal` has a cost over the overall performance of the
library. If it is not absolutely necessary for the calculations, it's
better not to use it.