Increase Ruby versions support
To make this library more accessible, I am increasing the number of
ruby versions supported.

I initially limited the support to very recent Ruby versions for two
reasons:
- support only branches in normal maintenance
- reduce the precision difference in BigDecimal with older versions

Now that BigDecimal has been removed from the project (e3d862f), this is
a good opportunity to increase the Ruby versions support, so that older
versions of Ruby are supported even if these versions are in EOL
maintenance mode.