Add twilight times
Due to the atmospheric refraction, the sky remains bright for some time
before the sunset and after the sunrise. This brightness of the sky is
called twilight and decreases proportionally to the Sun's angular
distance below the observer's horizon.

Between the day and the night, three different twilights are defined:
* Civil twilight: when the Sun's between 0° and 6° below the horizon
* Nautical twilight: when the Sun's between 6° and 12° below the horizon
* Astronomical twilight: when the Sun's between 12° and 18° below the
  horizon

This change introduces 6 new methods to `Astronoby::Sun` to provide the
instant in time (UTC) of each of these twilight changes:
* `#morning_civil_twilight_time`: when the morning civil twilight starts
* `#evening_civil_twilight_time`: when the evening civil twilight ends
* `#morning_nautical_twilight_time`: when the morning nautical twilight
  starts
* `#evening_nautical_twilight_time`: when the evening nautical twilight
  ends
* `#morning_astronomical_twilight_time`: when the night ends and the
  morning astronomical twilight starts
* `#evening_astronomical_twilight_time`: when the evening astronomical
  twilight ends and the night starts

On latitudes close to the poles, the night never completely falls for
some days in the year. When this happens, these methods will return
`nil` instead of a time.