Fix ecliptic to equatorial epoch
Before, when a set of ecliptic coordinates was converted to equatorial
ones, the given epoch was broken.

Now, it is possible to convert properly ecliptic coordinates for an
epoch, to equatorial coordinates for another epoch.

It is possible that this work will be improved with a more accurate
algorithm from jean Meeus, in the future.