Twilight events dedicated class
In the same way `Astronoby::Events::ObservationEvents` represent
observation events for any body on the celestial sphere, twilight events
exist but are related only to the Sun. For example, it doesn't make
sense to talk about twilight related to a planet or a distant star.

In a goal of making events more centralized, twilight times are now
extract from `Astronoby::Sun` into their own dedicated
`Astronoby::Events::TwilightEvents` class.

The same methods are exposed. This adds natural memoization as all the
times are calculated when the object initialized, which is when the
`Astronoby::Sun#twilight_events` is called.