Improve accuracy of equation of time
This new algorithm, based on _Astronomical Algorithms_ by Jean Meeus,
provide a more accurate algorithm for the equation of time than the
previous one. It also supports giving a precise time, or a date.

If a date is provided, the instant of time selected is 12PM UTC of the
date, which mostly explains the difference with official results, that
are often based only on a date (and an unknown instant of the day).