Expand the number of tested Ruby versions
Previously, only the latest patch version of each minor version since
Ruby 3.0.0 were tested.

Now, every patch version since 3.0.0 is tested to expand visibility and
insurance over all maintained MRI versions and beyond.

When this PR was created, Ruby 3.0 just got out of the EOL ([ref]). But
in order to be conservative, all patch versions of 3.0 are still tested,
and the minimum required Ruby version for the library remains 3.0.0.

[ref]: https://www.ruby-lang.org/en/downloads/branches/