Introduce observation events for Moon
Like the Sun, the Moon is a candidate for events defined in
`Astronoby::Events::ObservationEvents`, such as rising, transit and
setting times, or rising and setting azimuths, or transit altitude.

This adds `Astronoby::Moon#observation_events` method, with an
`observer` as parameter.

This change also refactors how observation events are computed,
introducing a loop to increase precision, as explained in _Astronomical
Algorithms_.