Fix `ObservationEvents` infinite loop
In some cases new iterations of rise-transit-set calculations are not
more precise than the previous ones. Quite the opposite, it is
apparently possible to loop indefinitely over some values without ever
crossing the targeted precision.

This introduces a limit of 5 iterations. Passed this limit, it is
admitted that the program in its current form will not achieve a better
precision.

Fixes #109