Add `Instant` value object (#121)
Time can be an ambiguous concept in Ruby and in astronomy.

In Astronoby, we often have to convert time into Julian Date, into Terrestrial Time, make mathematical operations on it, etc.

This introduces `Instant`, a value object that stores an instant in time in terrestrial time as a Julian Date. It allows for conversion into `Time`, `DateTime`, `Date`, but also TAI, GMST, get ΔT, get the difference with UTC in Julian Date.

It also supports comparison with other instances, and to get the difference in Julian Date with another instant.