Introduce `Astronoby::Ephem`
The new class is a wrapper around the Ephem gem.

It allows to download an ephemeris from the internet and load it using
the Ephem gem.

Unlike Ephem, Astronoby deals with ephemeris file in an agnostic way,
regardless of the format. Internally, it knows we're only dealing with
SPK files from NASA JPL, but this is not a constraint Astronoby users
should worry about.