Introduce Astrometric position for planets (#129)
In `Astronoby`, an astrometric position is a geocentric position in the International Celestial Reference Frame (ICRF) with light-time delay correction.

In other libraries, an astrometric position doesn't necessarily have to be geocentric, it can be with any center as long as it takes into account the light-time delay. For `Astronoby`, because the current focus is observation from Earth, the center is automatically the Earth center of mass.

While it includes a correction for the time the light takes to reach the center, it is not an "apparent" position, meaning it doesn't exactly describe where an object would be seen in the sky from an observer on Earth, as many other corrections must be applied to achieve this. But this is not the purpose of the astrometric position.

From the astrometric position, we will be able to compute the apparent position, but also to change frame from equatorial to ecliptic for some of the algorithms.

At the moment, this position is only available to the 8 planets of the Solar System.

This PR also adds `Distance#abs2` to be able to compute a vector magnitude, necessary for calculating the distance between the center and the target.

Because it is a geocentric position, the Earth astrometric position in `Astronoby` is a null vector.

As documented is specs, the accuracy is very high: right ascension and declination are matching the IMCCE's results below a tenth of arcsecond.

### Usage

```rb
time = Time.utc(2025, 4, 1)
instant = Astronoby::Instant.from_time(time)
ephem = Astronoby::Ephem.load("de440s.bsp")

mars = Astronoby::Mars.new(instant: instant, ephem: ephem)
astrometric = mars.astrometric

astrometric.right_ascension.str(:hms)
# => "7h 42m 19.2644s"

astrometric.declination.str(:dms)
# => "+24° 3′ 35.8686″"

astrometric.distance.kilometers.round
# => 170390364

astrometric.position.map(&:km).map(&:round)
# => Vector[-67178482, 140336314, 69466874]

astrometric.velocity.map(&:kmps).map { _1.round(2) }
# => Vector[-16.79, 9.77, 4.11]
```