Rename Barycentric into Geometric (#130)
This makes more sense as "barycentric" refers to the center of origin of
the reference frame, while "geometric" is really the term used to define
a position that hasn't been altered by any correction. It is a purely
geometrical position as in real physics light takes time to go from the
target to the center, the position described in theoretical.

If if in `Astronoby` a `Geometric` position is technically a barycentric
one.