Geometric and Astrometric reference frames with coordinates
This changes `Geometric` and `Astrometric` not as `Position` anymore but
`ReferenceFrame`.

The terminology is important `Geometric` and `Astrometric` don't really
describe a position by themselves, but a frame where the position is
defined.

Now, a `ReferenceFrame` exposes coordinates, with equatorial coordinates
being the first supported kind of coordinates. Ecliptic coordinates will
come soon.

A `ReferenceFrame` still exposes the `position`, `velocity` and
`distance`. It makes sense that it is not called a `Position` anymore as
the `position` is one attribute accessible from the reference frame.