Increase precision of mean obliquity (#133)
This updates the coefficients used to compute the mean obliquity to match P03 astronomical model.

The change is really minimal (less than 0.1 arcsecond) but the goal is to be as accurate as possible.

\<teasing>This change will make more sense in the close future when we'll update the nutation coefficients using IERS data.\</teasing>

I realise how inconvenient it is to have so much test changes, this is not ideal. This is something I would like to improve in the future, so that every tiny precision increase doesn't lead to updating half of the specs.
In the meantime, that's what's happening 🤦