Precession matrix for 2006 P03 model
This enables to generate a rotation matrix to apply the 2006 P03 astronomical model for precession.

Also called "precession of equinoxes", precession is a change of the orientation of the Earth's rotational axis. It is a large but slow shift of the rotation axis which changes the direction of the celestial poles.