New nutation model (#141)
This introduces `Astronoby::Nutation2`, temporarily named this way, as the new nutation model for Astronoby.

The astronomical nutation is a phenomenon which causes the orientation of the axis of rotation of a spinning astronomical object to vary over time.

It contributes to about 2% of the total precession, the long-period rotation of the Earth's rotation axis.

`Astronoby::Nutation2` relies on official terms from the IAU 2000B model, which is a simplified but still extremely accurate version of the IAU 2000A model.

IAU2000B contains the 77 most significant terms from the 1365 terms from IAU2000A, while it only sacrifices 1 milliarcsecond of accuracy throughout 1995-2025.

The matrix product of the precession matrix and the nutation matrix accounts for the total precession effect. This is necessary to compute the apparent position of a celestial object as seen from an observer on Earth.