Light deflection correction (#142)
This introduces `Astronoby::Deflection`, responsible for correcting an
astrometric position with the light deflection effect.

When light comes close to a massive body while on its way to an
observer, the light slightly bended due to relativistic effects.

In theory, all massive objects of the Solar System should be accounted
for light deflection. However, even the Sun itself has very little
effect, which is why this algorithm ignores light deflection except for
the Sun.

This algorithm is heavily inspired by Skyfield's algorithm of the same
effect.