Improve Vector integration with value objects
In Astronoby we have a custom `Astronoby::Vector` object that helps
manipulating value objects while still benefiting from the native `Vector`
Ruby API.

So far however the integration wasn't so great with the need to
frequently and manually loop over arrays to transform them into vectors
of value object.

This is now improved with the ability to create a vector of `Distance`
or `Velocity` objects from an array in the specified unit.

So far, only distances in meters and velocities in meters per second are
supported as they are the only unit necessary in the code base at the
moment.

`#magnitude` is also improved to return `Distance` or `Velocity` object
depending on what the vector is made of, without breaking the ability to
compute the magnitude from a vector of standard numeric values.