Handle refracted topocentric horizontal coordinates (#147)
By default, the topocentric `horizontal` coordinates are computed in a
theoretical air-less world, the position of an object in the sky as seen
from an observer if the atmosphere had no impact.

But the atmosphere does have an impact, called the atmospheric
refraction.

This enables to compute the horizontal coordinates using the `refraction`
keyword (set to `false` by default) to compute the refracted
coordinates.