Clean code after Ephem refactoring (#152)
After weeks of development of the new API based on the computed positions from ephemerides, it is time to clean the code, get rid of what not used anymore.

This is done with a heavy heart as it took me a very long time to write all the lines of code that are deleted today. This is the cost of moving Astronoby into a next chapter.