Improve `RisingTransitSettingEventsCalculator` (#155)
This includes some code organization changes, new specs and naming
changes, to better fit with my coding style.

For mathematical functions moved to `Util::Maths`, I was helped with an
AI model to write the specs. I understand the logic of the methods, but
I'm not strong enough in math to properly understand how to test edge
cases.