Lazy-load reference frames (#157)
Before, `geometric`, `astrometric`, `mean_of_date` and `apparent`
reference frames were computed immediately at the body's initialization.
This was due to the fact that they all need at least some piece of
information from other reference frames. This was also to make sure
`ephem` was never saved as instance variable as it can be a very large
one.

Now, only the required pieces of information are computed on
initialization:
* `geometric` reference frame
* Earth's geometric position
* Light-time correction

From this data, `astrometric`, `mean_of_date` and `apparent` positions
can be computed just-in-time. `ephem` is still not saved as instance
variable as it is not necessary anymore once the required data is
computed.

While this is not going to change much for parts of the library that use
the apparent position as it requires `geometric` and `astrometric` to be
computed, it is going to improve the overall performance of the library
by preventing to compute all reference frames when they're not all
necessary.