Remove CodeCov integration (#162)
The tool is not behaving the way I want and keep messing with my PRs even with multiple attempts at editing the configuration.

I know the problem comes from me and I'm just not doing things right, but it's taking more time than I wanted to spend on it and I don't like the impact it has on the developer experience at the moment, so I'm removing it.

Until I retry to use CodeCov, the test coverage is still accessible locally or in the CI logs.