Add support for IMCCE INPOP
_INPOP_ is the ephemerides series from the Paris Observatory, more
precisely from the IMCCE laboratory. The Paris Observatory has been a
central place for astronomy for the last few centuries and the data
produces today by the IMCCE is among the most accurate and precise we
can produce today.IMCCE

It is now possible to load Binary SPICE files (.bsp) produced by the
IMCCE for their INPOP model.

Co-authored-by: Joël Quenneville <joelq@thoughtbot.com>