Drop `Astronoby::Observer#observe`
`#observe` was in fact broken as it computes a topocentric position from
an apparent one, but the method doesn't have what is needed for computed
an apparent position.

The method could be fixed, but because the same result could be
obtained using `Astronoby::SolarSystemBody#observed_by` I don't see it
necessary to have two methods in the end.