Introduce performance benchmark (#183)
There will always be room for improvement regarding performance, but it's hard to know what and how to measure during development phase.

This adds a performance benchmark that can be executed locally in order to provide feedback on the impact on performance of a new change.

This first version benchmarks the main expensive features of the library. The goal is not to stick to absolute results, as they will depend on the computer running the benchmark. The correct use is to compare the results from a branch with the results on `main`.

### Usage

```
$ bin/console
```

```rb
require_relative "benchmarks/performance"

PerformanceBenchmark.new.run
```

For fast feedback, you can customise the parameters:

```rb
PerformanceBenchmark
  .new(warmup_runs: 1, measure_runs: 3, iterations_per_run: 5)
  .run
```

Just as an indication of the result format, here is the output from  my computer:

```
Astronoby Performance Benchmark
  Ruby version : 3.4.3
  Date         : 2025-05-20

Warming up (2 runs, not measured)...

Measuring (3 runs, each 10 iterations)...
  Run 1/3 complete.
  Run 2/3 complete.
  Run 3/3 complete.

RESULTS (averaged over 3 runs, each 10 iterations):
  rts_event_on:          26.66  ±  0.18 sec
  rts_events_between:     22.47  ±  0.18 sec
  twilight_event_on:     27.46  ±  0.18 sec
  moon_phases:           22.19  ±  0.12 sec
  total:                 98.78  ±  0.66 sec
```