Cache positions in RTS calculator (#182)
When computing the rise, transit and set times of a celestial body, the position of the body could be computed multiple times for the same instant. This is not optimized, especially while computing a position is one of the most expensive operation of the library, as we speak.

This caches the computed position during a single operation (`event_on`, ...).

### Comparison

#### Initial

```
RESULTS (averaged over 3 runs, each 10 iterations):
  rts_event_on:          26.66  ±  0.18 sec
  rts_events_between:     22.47  ±  0.18 sec
  twilight_event_on:     27.46  ±  0.18 sec
  moon_phases:           22.19  ±  0.12 sec
  total:                 98.78  ±  0.66 sec
```

#### Fixed

```
RESULTS (averaged over 3 runs, each 10 iterations):
  rts_event_on:          17.11  ±  0.06 sec
  rts_events_between:     13.16  ±  0.03 sec
  twilight_event_on:     18.89  ±  0.05 sec
  moon_phases:           22.24  ±  0.06 sec
  total:                 71.40  ±  0.17 sec
```

#### Conclusion

* -36% for `Astronoby::RiseTransitSetCalculator#event_on`
* -41% for `Astronoby::RiseTransitSetCalculator#events_between`
* -31% for `Astronoby::TwilightCalculator#event_on`
* (no change for `Astronoby::Events::MoonPhases`)