Global configuration (#187)
With #186, global cache was introduced to increase the performance of the library on the main features. However, because this caching implementation goes with a small reduction of precision (instants are rounded to increase the chance of hitting it), it felt odd to force the user to use it, and to use it exclusively of the author decided.

This introduces a global configuration with a nice API to let the user customise it.

## Usage

Currently, the configuration has two purposes:
* enabling/disabling caching
* set customised rounding values for the features using cache

```rb
Astronoby.configuration.cache_enabled = true
Astronoby.cache_precision(:geometric, 5)
```

This will enable caching globally and geometric positions will be cached based on a rounding of 0.00001 on the instant. As the instant is stored as a Julian Day, this represents 0.1 seconds: the same geometric position won't be re-computed if its instant falls into 0.1 seconds of a previously computed one.

## Defaults

By default, caching is disabled. I want to avoid new users discovering the library and having unexpected results. The default behaviour is to compute everything from scratch constantly, except positions in the RTS calculator during computation.

## Performance

With no cache by default, the performance is back to a few commits before, even slightly less performant due to the new added complexity. However, with cache enabled, we're back to interesting performance, with the added bonus of customisable configuration.

```rb
require_relative "benchmarks/performance"
PerformanceBenchmark.new.run

# Astronoby Performance Benchmark
#   Ruby version : 3.4.3
#   Date         : 2025-05-26
#
# Warming up (2 runs, not measured)...
#
# Measuring (3 runs, each 10 iterations)...
#   Run 1/3 complete.
#   Run 2/3 complete.
#   Run 3/3 complete.
#
# RESULTS (averaged over 3 runs, each 10 iterations):
#   rts_event_on:          22.20  ±  0.06 sec
#   rts_events_between:     17.23  ±  0.05 sec
#   twilight_event_on:     23.83  ±  0.30 sec
#   moon_phases:           22.66  ±  0.06 sec
#   total:                 85.92  ±  0.32 sec
```

```rb
Astronoby.configuration.cache_enabled = true
require_relative "benchmarks/performance"
PerformanceBenchmark.new.run

# Astronoby Performance Benchmark
#   Ruby version : 3.4.3
#   Date         : 2025-05-26
#
# Warming up (2 runs, not measured)...
#
# Measuring (3 runs, each 10 iterations)...
#   Run 1/3 complete.
#   Run 2/3 complete.
#   Run 3/3 complete.
#
# RESULTS (averaged over 3 runs, each 10 iterations):
#   rts_event_on:          10.23  ±  0.04 sec
#   rts_events_between:      8.60  ±  0.07 sec
#   twilight_event_on:      8.18  ±  0.04 sec
#   moon_phases:            0.06  ±  0.00 sec
#   total:                 27.06  ±  0.06 sec
```