`Epoch` refactoring into `JulianDate` (#197)
Terms were confusing when dealing with an `epoch`. An epoch is a reference point, but it depends on a system used. In Astronoby, the system used is the Julian Day, so it wasn't precise enough to call it `Epoch`.

Also, specs where mentioning "Julian Day Number", which is the integer part of a Julian Date, which was not accurate.

This refactors `Epoch` into `JulianDate` for accuracy.

This also drops `#to_utc` method, which ambiguously expected the Julian date to be in UTC, which is almost never the case in Astronoby (`Instant` used Julian dates, but stores the value in TT). The code using it has been adapted to use `Instant.from_terrestrial_time` instead, which has increased precision by a couple seconds in the equinox/solstice calculation.

It was also the opportunity to add `#from_julian_year` and `#from_besselian_year` to compute Julian dates from reference points like `J2050` or `B1875`. This will help introduce constellations, which boundaries are defined based on the `B1875` epoch.